BAEL-3120 Counting Words in a String
