--- spring-boot-data/src/main/java/com/baeldung/hexagonal/domain/Product.java
@@ -1,150 +1,0 @@
-package com.baeldung.hexagonal.domain;
-
-import java.io.Serializable;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import javax.persistence.Version;
-import javax.xml.bind.annotation.XmlRootElement;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-
-/**
- * 
- * @author shwetaJ Model & API Object for Product
- *
- */
-
-@XmlRootElement
-@Entity
-@Table(name = "PRODUCT")
-public class Product implements Serializable {
-	private static final long serialVersionUID = 1L;
-
-	@Id
-	@GeneratedValue
-	@Column(name = "PRODUCT_ID")
-	private int productId;
-
-	@Column(name = "PRODUCT_NAME")
-	private String productName;
-
-	@Column(name = "PRODUCT_DESC")
-	private String productDesc;
-
-	@Column(name = "PRODUCT_CATEGORY")
-	private String productCategory;
-
-	@Column(name = "PRODUCT_PRICE")
-	private int productPrice;
-
-	@JsonIgnore
-	@Version
-	@Column(name = "VERSION")
-	private int version;
-
-	/**
-	 * @return the productId
-	 */
-	public int getProductId() {
-		return productId;
-	}
-
-	/**
-	 * @param productId
-	 *            the productId to set
-	 */
-	public void setProductId(int productId) {
-		this.productId = productId;
-	}
-
-	/**
-	 * @return the productName
-	 */
-	public String getProductName() {
-		return productName;
-	}
-
-	/**
-	 * @param productName
-	 *            the productName to set
-	 */
-	public void setProductName(String productName) {
-		this.productName = productName;
-	}
-
-	/**
-	 * @return the productDesc
-	 */
-	public String getProductDesc() {
-		return productDesc;
-	}
-
-	/**
-	 * @param productDesc
-	 *            the productDesc to set
-	 */
-	public void setProductDesc(String productDesc) {
-		this.productDesc = productDesc;
-	}
-
-	/**
-	 * @return the productCategory
-	 */
-	public String getProductCategory() {
-		return productCategory;
-	}
-
-	/**
-	 * @param productCategory
-	 *            the productCategory to set
-	 */
-	public void setProductCategory(String productCategory) {
-		this.productCategory = productCategory;
-	}
-
-	/**
-	 * @return the productPrice
-	 */
-	public int getProductPrice() {
-		return productPrice;
-	}
-
-	/**
-	 * @param productPrice
-	 *            the productPrice to set
-	 */
-	public void setProductPrice(int productPrice) {
-		this.productPrice = productPrice;
-	}
-
-	/**
-	 * @return the version
-	 */
-	public int getVersion() {
-		return version;
-	}
-
-	/**
-	 * @param version
-	 *            the version to set
-	 */
-	public void setVersion(int version) {
-		this.version = version;
-	}
-
-	/*
-	 * (non-Javadoc)
-	 * 
-	 * @see java.lang.Object#toString()
-	 */
-	@Override
-	public String toString() {
-		return "Product [productId=" + productId + ", productName=" + productName + ", productDesc=" + productDesc
-				+ ", productCategory=" + productCategory + ", productPrice=" + productPrice + ", version=" + version
-				+ "]";
-	}
-}
