--- spring-boot-data/src/main/java/com/baeldung/hexagonal/service/ProductService.java
@@ -1,18 +1,0 @@
-package com.baeldung.hexagonal.service;
-
-import java.util.List;
-
-import com.baeldung.hexagonal.domain.Product;
-
-/**
- * 
- * @author shwetaJ
- *
- */
-
-public interface ProductService {
-	
-	Product createProduct(Product product);
-	Product getProduct(int productId);
-	List<Product> getAllProducts();
-}
